Final Project Wireframe Sketches and Description
My final project will be a website for a good friend of mine from the graduate program I dropped out of earlier this year. She will be graduating with her Masters in Marriage and Family Therapy, with a few other unique certifications, and wants her professional independent private practice website to be representative of her specialities (LGBT Issues, Military Families, Addiction) and reflect her individuality. 

It will be a single page website, with a sticky header (which contains the navigation with jump links to specific sections down the page) and sticky footer containing a button which takes you directly to the appointment request form section at the very bottom of the page.

Ideally, all sections will contain interactive elements, ie. hover effects, overlays, etc. I've detailed those specifics in the sketches. My friend has been so kind as to write out some text for portions of the site, like responses to questions in the FAQ, and her statement regarding LGBT+ Affirming Experiential Approach.